Change int to long in an_model c conde for compatibility

+++ cnmodel/an_model/model/model_IHC.c
@@ -60,7 +60,7 @@ void mexFunction(int nlhs, mxArray *plhs[], int nrhs, const mxArray *prhs[])
 {
 	
 	double *px, cf, tdres, reptime, cohc, cihc;
-	int    nrep, pxbins, lp, outsize[2], totalstim, species;
+	long    nrep, pxbins, lp, outsize[2], totalstim, species;
 
 	double *pxtmp, *cftmp, *nreptmp, *tdrestmp, *reptimetmp, *cohctmp, *cihctmp, *speciestmp;
     double *ihcout;
